dev_utils/dev_lab/geoserver-rest/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
GDAL==3.6.2
geoserver-rest==2.3.4
idna==3.4
Js2Py==0.74
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycurl==7.45.2
Pygments==2.14.0
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
xmltodict==0.13.0
